+++ agones/Cargo.toml
@@ -25,9 +25,10 @@ readme = "README.md"
 
 [dependencies]
 base64.workspace = true
+futures.workspace = true
 k8s-openapi.workspace = true
 kube = { workspace = true, features = ["openssl-tls", "client", "derive", "runtime"] }
 quilkin = { path = "../" }
+serial_test = "2.0.0"
 tokio.workspace = true
-futures.workspace = true
 tracing.workspace = true
